Store
Throughout this file, you can notice that I keep track of the sales (with an ID, sale price, date and quantity of products sold) and purchases (with an ID, purchase price, date and quantity of products bought). There's a dictionary that  updates itself as you buy and sell, making a plausible model. All of this to better understand and get statistics that would later on benefit the company.